Fix constructor and export types
Constructor functions return a value that doesn't match the exported type. This means that the World constructor is returning a value that matches _World and is trying to cast it to World which is the public type that's being exported. Fix these problems by tweaking in some way the types...

Also move Component type from being exported directly from Component.luau to being part of Types.luau. The type doesn't require code present in Component.luau to exist and is more convenient to handle with Types.luau.